Debugged `ConditionalPlanarFlow`
