+++ pyro/distributions/__init__.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import pyro.distributions.torch_patch  # noqa F403
 from pyro.distributions.avf_mvn import AVFMultivariateNormal
+from pyro.distributions.conditional_transform import ConditionalTransform
 from pyro.distributions.conditional_transformed_distribution import ConditionalTransformedDistribution
 from pyro.distributions.conjugate import BetaBinomial, DirichletMultinomial, GammaPoisson
 from pyro.distributions.delta import Delta
@@ -31,6 +32,7 @@ from pyro.distributions.transforms import * # noqa F403
 __all__ = [
     "AVFMultivariateNormal",
     "BetaBinomial",
+    "ConditionalTransform",
     "ConditionalTransformedDistribution",
     "ConditionalTransformModule",
     "Delta",
