+++ pyro/distributions/transforms/householder.py
@@ -7,7 +7,7 @@ import torch
 import torch.nn as nn
 from torch.distributions import constraints
 
-from pyro.distributions.torch_transform import TransformModule
+from pyro.distributions import TransformModule
 from pyro.distributions.util import copy_docs_from
