+++ pyro/distributions/transforms/permute.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
 import torch
-from torch.distributions.transforms import Transform
+from torch.distributions import Transform
 from torch.distributions.utils import lazy_property
 from torch.distributions import constraints
